Move InterpretertState to state.py.
